Apps crashing when disclaimer is not cancelled or accepted (#2923)
* wait for disclaimer accept before initializing app

* Broken apps not working -> Add apploadError to the condition for not executing the timers

* error boundary not being hit because throwing in async fn inside effect

* Manage LegalDisplaymer in AppFrame parent

* Import relative to src
