+++ src/routes/safe/components/Apps/components/AppFrame.tsx
@@ -10,21 +10,19 @@ import {
   SignMessageParams,
   RequestId,
 } from '@gnosis.pm/safe-apps-sdk'
-import { useHistory } from 'react-router-dom'
+
 import { useSelector } from 'react-redux'
 import { INTERFACE_MESSAGES, Transaction, LowercaseNetworks } from '@gnosis.pm/safe-apps-sdk-v1'
 import Web3 from 'web3'
 
 import { currentSafe } from 'src/logic/safe/store/selectors'
-import { getCurrentShortChainName, getNetworkName, getSafeAppsRpcServiceUrl, getTxServiceUrl } from 'src/config'
+import { getNetworkName, getSafeAppsRpcServiceUrl, getTxServiceUrl } from 'src/config'
 import { isSameURL } from 'src/utils/url'
 import { useAnalytics, SAFE_NAVIGATION_EVENT } from 'src/utils/googleAnalytics'
 import { LoadingContainer } from 'src/components/LoaderContainer/index'
 import { TIMEOUT } from 'src/utils/constants'
 import { ConfirmTxModal } from './ConfirmTxModal'
 import { useIframeMessageHandler } from '../hooks/useIframeMessageHandler'
-import { useLegalConsent } from '../hooks/useLegalConsent'
-import LegalDisclaimer from './LegalDisclaimer'
 import { getAppInfoFromUrl } from '../utils'
 import { SafeApp } from '../types'
 import { useAppCommunicator } from '../communicator'
@@ -35,7 +33,6 @@ import { addressBookEntryName } from 'src/logic/addressBook/store/selectors'
 import { currentChainId } from 'src/logic/config/store/selectors'
 import { useSignMessageModal } from '../hooks/useSignMessageModal'
 import { SignMessageModal } from './SignMessageModal'
-import { generateSafeRoute, SAFE_ROUTES } from 'src/routes/routes'
 
 const AppWrapper = styled.div`
   display: flex;
@@ -86,28 +83,24 @@ const safeAppWeb3Provider = new Web3.providers.HttpProvider(getSafeAppsRpcServic
   timeout: 10_000,
 })
 
+const URL_NOT_PROVIDED_ERROR = 'App url No provided or it is invalid.'
+const APP_LOAD_ERROR = 'There was an error loading the Safe App. There might be a problem with the App provider.'
+
 const AppFrame = ({ appUrl }: Props): ReactElement => {
   const { address: safeAddress, ethBalance, owners, threshold } = useSelector(currentSafe)
   const networkId = useSelector(currentChainId)
   const safeName = useSelector((state) => addressBookEntryName(state, { address: safeAddress }))
   const { trackEvent } = useAnalytics()
-  const history = useHistory()
-  const { consentReceived, onConsentReceipt } = useLegalConsent()
-
   const iframeRef = useRef<HTMLIFrameElement>(null)
   const [confirmTransactionModal, setConfirmTransactionModal] =
     useState<ConfirmTransactionModalState>(INITIAL_CONFIRM_TX_MODAL_STATE)
   const [appIsLoading, setAppIsLoading] = useState<boolean>(true)
   const [safeApp, setSafeApp] = useState<SafeApp | undefined>()
   const [signMessageModalState, openSignMessageModal, closeSignMessageModal] = useSignMessageModal()
-
-  const redirectToBalance = () =>
-    history.push(generateSafeRoute(SAFE_ROUTES.ASSETS_BALANCES, { shortName: getCurrentShortChainName(), safeAddress }))
   const timer = useRef<number>()
   const [isLoadingSlow, setIsLoadingSlow] = useState<boolean>(false)
-
   const errorTimer = useRef<number>()
-  const [appLoadError, setAppLoadError] = useState<boolean>(false)
+  const [, setAppLoadError] = useState<boolean>(false)
 
   useEffect(() => {
     const clearTimeouts = () => {
@@ -120,12 +113,13 @@ const AppFrame = ({ appUrl }: Props): ReactElement => {
         setIsLoadingSlow(true)
       }, TIMEOUT)
       errorTimer.current = window.setTimeout(() => {
-        setAppLoadError(true)
+        setAppLoadError(() => {
+          throw Error(APP_LOAD_ERROR)
+        })
       }, APP_LOAD_ERROR_TIMEOUT)
     } else {
       clearTimeouts()
       setIsLoadingSlow(false)
-      setAppLoadError(false)
     }
 
     return () => {
@@ -266,6 +260,10 @@ const AppFrame = ({ appUrl }: Props): ReactElement => {
   }
 
   useEffect(() => {
+    if (!appUrl) {
+      throw Error(URL_NOT_PROVIDED_ERROR)
+    }
+
     const loadApp = async () => {
       try {
         const app = await getAppInfoFromUrl(appUrl)
@@ -274,6 +272,7 @@ const AppFrame = ({ appUrl }: Props): ReactElement => {
         logError(Errors._900, `${appUrl}, ${err.message}`)
       }
     }
+
     loadApp()
   }, [appUrl])
 
@@ -283,14 +282,6 @@ const AppFrame = ({ appUrl }: Props): ReactElement => {
       trackEvent({ category: SAFE_NAVIGATION_EVENT, action: 'Apps', label: safeApp.name })
     }
   }, [safeApp, trackEvent])
-
-  if (!appUrl) {
-    throw Error('App url No provided or it is invalid.')
-  }
-
-  if (appLoadError) {
-    throw Error('There was an error loading the Safe App. There might be a problem with the App provider.')
-  }
 
   if (!safeApp) {
     return (
@@ -303,10 +294,6 @@ const AppFrame = ({ appUrl }: Props): ReactElement => {
         <Loader size="md" />
       </LoadingContainer>
     )
-  }
-
-  if (consentReceived === false) {
-    return <LegalDisclaimer onCancel={redirectToBalance} onConfirm={onConsentReceipt} />
   }
 
   return (
